Minor doc cleanup, component rename
