+++ docusaurus/docs/Android/04-compose/03-channel-components/01-channels-screen.mdx
@@ -2,14 +2,16 @@
 
 The easiest way to set up a screen that shows the active user's channels and gives them the ability to search for a specific channel is to use the `ChannelsScreen`.
 
-`ChannelsScreen` sets up four components internally:
+`ChannelsScreen` sets up the following functionality internally:
 
-* [`ChannelListHeader`](./02-channel-list-header.mdx): Shows the information of the current user and exposes a trailing action.
-* [`SearchInput`](../06-utility-components/03-search-input.mdx): Allows users to search for channels by name.
-* [`ChannelList`](./03-channel-list.mdx): The core part of the screen, which renders channels for the active user, based on defined filters.
-* [`ChannelInfo`](./05-channel-info.mdx): A bottom drawer that is shown when long clicking on a `Channel` in the list.
+* Header with the information of the current user and customizable title and action.
+* Search input that can be shown or hidden that allows users to search for channels by name.
+* List of user's channels with pagination, based on defined filters and sorting.
+* Menu with detailed channel information, shown when long clicking on a `Channel` in the list.
 
 It also sets up all the business logic and styles the UI according to our default design system.
+
+Let's see how to integrate the component.
 
 ## Usage
 
@@ -85,9 +87,9 @@ override fun onCreate(savedInstanceState: Bundle?) {
 
 ## Customization
 
-`ChannelsScreen` is one of our **screen components** and as such, it doesn't offer much customization. As with any component, you can customize the content theme and styling by wrapping it the [`ChatTheme`](../02-general-customization/01-chat-theme.mdx).
+`ChannelsScreen` is one of our **screen components** and as such it doesn't offer much customization. As with any component, you can customize the content theme and styling by wrapping it the [`ChatTheme`](../02-general-customization/01-chat-theme.mdx).
 
-When it comes to UI and behavior customization in the `ChannelsScreen` signature, you have access to the following:
+When it comes to UI and behavior customization in the `ChannelsScreen` signature you have access to the following:
 
 ```kotlin
 fun ChannelsScreen(
@@ -107,6 +109,6 @@ fun ChannelsScreen(
 * `querySort`: Like with filters, these are applied to the list data and affect its sorting order.
 * `title`: The title of the `ChannelListHeader`.
 * `isShowingHeader`: Flag that affects if we show the `ChannelListHeader`. `true` by default.
-* `isShowingSearch`: Flag that affects if we show the `SearchInput`. `true` by default.
+* `isShowingSearch`: Flag that affects if we show the `SearchInput`. `false` by default.
 
 <!-- TODO WIP PAGE If you want to build a custom Channels screen UI or override the default the behavior, follow our [Building Custom Screens](../_07-guides/06-building-custom-screens.mdx) guide. -->
